% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{rc-reference-medians}
\alias{rc-reference-medians}
\title{Recentering reference medians}
\description{
Contains reference median values for default recentering in the child
algorithm
}
\section{rcfile-resmoothed.csv.gz}{


Used in function \code{cleangrowth()}
}

